List and Export a WiFi Profile
Batchfile queries the host computer for a list of known WiFi network profiles.  The user is then asked to provide the name of a WiFi network they wish to export.  From there an XML file is placed on the 
desktop and can be copied to another computer.  When a script on the target machine is executed the profile information can be imported without the need to manually connect the host computer the WiFi network.